# docs/swagger.yaml
definitions:
  request.GeminiHistory:
    properties:
      answer:
        type: string
      question:
        type: string
    required:
    - answer
    - question
    type: object
  request.PromptText:
    properties:
      model:
        type: string
      text:
        type: string
    required:
    - text
    type: object
  request.UpdateGeminiHistory:
    properties:
      answer:
        type: string
      id:
        type: integer
      question:
        type: string
    required:
    - answer
    - id
    - question
    type: object
info:
  contact: {}
paths:
  /gemini-history/history-data:
    delete:
      description: delete history data
      parameters:
      - description: history ID to be deleted
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses: {}
      summary: DeleteHistoryData
      tags:
      - gemini-history
    get:
      description: get all history data
      produces:
      - application/json
      responses: {}
      summary: GetAllHistoryData
      tags:
      - gemini-history
    patch:
      description: update history data
      parameters:
      - description: Request Body
        in: body
        name: UpdateGeminiHistory
        required: true
        schema:
          $ref: '#/definitions/request.UpdateGeminiHistory'
      produces:
      - application/json
      responses: {}
      summary: UpdateHistoryData
      tags:
      - gemini-history
    post:
      description: create history data
      parameters:
      - description: Request Body
        in: body
        name: CreateGeminiHistory
        required: true
        schema:
          $ref: '#/definitions/request.GeminiHistory'
      produces:
      - application/json
      responses: {}
      summary: CreateHistoryData
      tags:
      - gemini-history
  /gemini-history/history-data/bulk:
    post:
      description: create many history data
      parameters:
      - description: Request Body
        in: body
        name: CreateManyGeminiHistory
        required: true
        schema:
          items:
            $ref: '#/definitions/request.GeminiHistory'
          type: array
      produces:
      - application/json
      responses: {}
      summary: CreateManyHistoryData
      tags:
      - gemini-history
  /gemini/chat:
    post:
      description: chat action API
      parameters:
      - description: Request Body
        in: body
        name: PromptText
        required: true
        schema:
          $ref: '#/definitions/request.PromptText'
      produces:
      - application/json
      responses: {}
      summary: Chat
      tags:
      - gemini
  /gemini/prompt-text:
    post:
      description: get prompt text result by prompt string
      parameters:
      - description: Request Body
        in: body
        name: PromptText
        required: true
        schema:
          $ref: '#/definitions/request.PromptText'
      produces:
      - application/json
      responses: {}
      summary: Prompt Text
      tags:
      - gemini
  /gemini/reset:
    get:
      description: for resetting all chat session
      produces:
      - application/json
      responses: {}
      summary: Reset chat session
      tags:
      - gemini
  /ollama/prompt-text:
    post:
      description: get Ollama prompt text result by defining the model and text
      parameters:
      - description: Request Body
        in: body
        name: PromptText
        required: true
        schema:
          $ref: '#/definitions/request.PromptText'
      produces:
      - application/json
      responses: {}
      summary: Prompt Ollama Text
      tags:
      - ollama
swagger: "2.0"